+++ packages/sns/candid/sns_governance_test.certified.idl.js
@@ -464,7 +464,7 @@ export const idlFactory = ({ IDL }) => {
     'topic_followees' : IDL.Opt(
       IDL.Record({
         'topic_id_to_followees' : IDL.Vec(
-          IDL.Tuple(IDL.Nat64, FolloweesForTopic)
+          IDL.Tuple(IDL.Int32, FolloweesForTopic)
         ),
       })
     ),
@@ -1262,7 +1262,7 @@ export const init = ({ IDL }) => {
     'topic_followees' : IDL.Opt(
       IDL.Record({
         'topic_id_to_followees' : IDL.Vec(
-          IDL.Tuple(IDL.Nat64, FolloweesForTopic)
+          IDL.Tuple(IDL.Int32, FolloweesForTopic)
         ),
       })
     ),