+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add `neuronMinimumDissolveDelayToVoteSeconds` to nns `VotingPowerEconomics`.
 - Expose method `status` in class `IcrcIndexNgCanister`.
 - Expose method `icrc3_get_blocks` in class `IcrcLedgerCanister`.
+- Add `setFollowing` to sns governance canister.
 - New service util `queryAndUpdate` that aids in executing a request and handling the results, for both type of calls.
 - Added type predicate to narrow value to `undefined`, `null`, or empty string in type guards in util `isEmptyString`.